requirements.txt
Django>=2
django-extensions>=2
djangorestframework==3.7.7
django-crispy-forms==1.7.0
psycopg2-binary==2.7.5
channels==1.1.8
gunicorn
whitenoise
django-heroku
coverage
django-coverage
celery==4.3.0
mistune==0.8.4
Pillow==6.0.0
xxhash==1.3.0
django-rest-auth==0.9.5
djangorestframework==3.7.7
django-model-utils==3.2.0
django-createsuperuserwithpassword
django-filter==2.2.0
django-addanother==2.0.0
textblob